std::find instead looking by the index
This was nasty and rogue bug that almost have driven me crazy.

The case was with looking for the item in collection of std::vector<PriorityQueueEntry *>
pointers and comparing it with the NULL which indeed absolutely legal and I think semantically
valid.

But instead it found nothing all the time and caused test to fail.

Use std::find if you need to look up for something in a collection of
container content.